+++ commons/src/test/java/org/fuin/srcgen4j/commons/GeneratedArtifactTest.java
@@ -47,17 +47,17 @@ public class GeneratedArtifactTest extends AbstractTest {
         final String name = "abc";
         final String pathAndName = "def";
         final byte[] source = "ghi".getBytes();
-        final String project = "prj";
+        final String module = "prj";
         final String folder = "fld";
 
         // TEST
-        final GeneratedArtifact testee = new GeneratedArtifact(name, pathAndName, source, project, folder);
+        final GeneratedArtifact testee = new GeneratedArtifact(name, pathAndName, source, module, folder);
 
         // VERIFY
         assertThat(testee.getName()).isEqualTo(name);
         assertThat(testee.getPathAndName()).isEqualTo(pathAndName);
         assertThat(testee.getData()).isEqualTo(source);
-        assertThat(testee.getProject()).isEqualTo(project);
+        assertThat(testee.getModule()).isEqualTo(module);
         assertThat(testee.getFolder()).isEqualTo(folder);
 
     }