+++ core/src/test/java/org/fuin/srcgen4j/core/emf/AbstractHelloTstGen.java
@@ -23,7 +23,7 @@ public final class AbstractHelloTstGen implements ArtifactFactory<Greeting> {
 
     private String artifact;
 
-    private String project;
+    private String module;
 
     private String folder;
 
@@ -39,7 +39,7 @@ public final class AbstractHelloTstGen implements ArtifactFactory<Greeting> {
     @Override
     public final void init(final ArtifactFactoryConfig config) {
         this.artifact = config.getArtifact();
-        this.project = config.getProject();
+        this.module = config.getModule();
         this.folder = config.getFolder();
         this.incremental = config.isIncremental();
         this.varMap = config.getVarMap();
@@ -61,7 +61,7 @@ public final class AbstractHelloTstGen implements ArtifactFactory<Greeting> {
             final String path = pkg.replace('.', '/');
             vars.put("package", pkg);
             return Collections.singletonList(new GeneratedArtifact(artifact, path + "/AbstractHello" + greeting.getName() + ".java",
-                    Utils4J.replaceVars(src, vars).getBytes(), project, folder));
+                    Utils4J.replaceVars(src, vars).getBytes(), module, folder));
         } catch (final IOException ex) {
             throw new RuntimeException(ex);
         }