+++ core/src/test/java/org/fuin/srcgen4j/core/emf/HelloPropertiesTstGen.java
@@ -20,7 +20,7 @@ public class HelloPropertiesTstGen implements ArtifactFactory<ResourceSet> {
 
     private String artifact;
 
-    private String project;
+    private String module;
 
     private String folder;
 
@@ -34,7 +34,7 @@ public class HelloPropertiesTstGen implements ArtifactFactory<ResourceSet> {
     @Override
     public final void init(final ArtifactFactoryConfig config) {
         this.artifact = config.getArtifact();
-        this.project = config.getProject();
+        this.module = config.getModule();
         this.folder = config.getFolder();
         this.incremental = config.isIncremental();
     }
@@ -59,7 +59,7 @@ public class HelloPropertiesTstGen implements ArtifactFactory<ResourceSet> {
             }
         }
 
-        return Collections.singletonList(new GeneratedArtifact(artifact, "hello.properties", sb.toString().getBytes(), project, folder));
+        return Collections.singletonList(new GeneratedArtifact(artifact, "hello.properties", sb.toString().getBytes(), module, folder));
 
     }